:recycle: refactor: rename security directory
